+++ src/main/java/Duke.java
@@ -1,3 +1,9 @@
+import duke.Parser;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.command.Command;
+import duke.exception.DukeException;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
