+++ src/main/java/duke/command/DeadlineCommand.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.exception.DukeException;
+import duke.task.Deadline;
+import duke.task.Task;
 import java.io.IOException;
 import java.time.DateTimeException;
