+++ src/main/java/duke/exception/EmptyFieldException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class EmptyFieldException extends DukeException {
 
     public EmptyFieldException(String message) {
