+++ src/main/java/duke/task/Event.java
@@ -1,23 +1,26 @@
+package duke.task;
+
+import duke.exception.EmptyFieldException;
+import duke.exception.InvalidCommandException;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import static java.time.temporal.ChronoUnit.MINUTES;
 import static java.time.temporal.ChronoUnit.DAYS;
 
+public class Event extends Task {
 
-public class Deadline extends Task {
-
-    protected String by;
+    protected String at;
     protected LocalDate date;
     protected LocalTime time;
 
-    public Deadline(String description, String by) throws EmptyFieldException, InvalidCommandException {
+    public Event(String description, String at) throws EmptyFieldException, InvalidCommandException {
         super(description);
-        if (description.equals("") || by.equals("")) {
+        if (description.equals("") || at.equals("")) {
             //check for empty description or by
             throw new EmptyFieldException("     Error! Description or date and time is empty!");
         } else {
-            String[] dateTime = by.split(" ");
+            String[] dateTime = at.split(" ");
             if (dateTime.length > 2) {
                 //check for more than 1 date and time entered
                 throw new InvalidCommandException(
@@ -25,7 +28,7 @@ public class Deadline extends Task {
                                 + "     one date, Eg: \"2021-09-12\" (This will enter time as 23:59 by default),\n"
                                 + "     or one time, Eg: \"18:00\" (This will enter today's date by default)");
             } else {
-                this.by = by;
+                this.at = at;
                 if (dateTime.length == 1) {
                     //user only entered a single date or time
                     if (dateTime[0].length() > 5) {
@@ -48,7 +51,7 @@ public class Deadline extends Task {
         }
     }
 
-    private void checkChronologicalOrder() throws InvalidCommandException{
+    private void checkChronologicalOrder() throws InvalidCommandException {
         LocalDate dateNow = LocalDate.now();
         LocalTime timeNow = LocalTime.now();
         if (dateNow.until(this.date, DAYS) < 0) {
@@ -65,17 +68,17 @@ public class Deadline extends Task {
         }
     }
 
-
     @Override
     public String saveToFile() {
-        return "D " + super.saveToFile() + " | " + this.by;
+        return "E " + super.saveToFile() + " | " + this.at;
     }
 
     @Override
     public String toString() {
-        return "[D]"
+        return "[E]"
                 + super.toString()
-                + " (by: " + this.date.format(DateTimeFormatter.ofPattern("MMMM d yyyy"))
+                + " (at: "
+                + this.date.format(DateTimeFormatter.ofPattern("MMMM d yyyy"))
                 + ", "
                 + this.time.format(DateTimeFormatter.ofPattern("h:mm a"))
                 + ")";
